Struct crossbeam::mem::epoch::garbage::Local

pub struct Local {
    pub old: Bag,
    pub cur: Bag,
    pub new: Bag,
}

A thread-local set of garbage bags.

Fields
old: Bag

Garbage added at least one epoch behind the current local epoch

cur: Bag

Garbage added in the current local epoch or earlier

new: Bag

Garbage added in the current global epoch

Methods

impl Local

fn new() -> Local

fn insert<T>(&mut self, elem: *mut T)

unsafe fn collect(&mut self)

Collect one epoch of garbage, rotating the local garbage bags.

fn size(&self) -> usize