Struct crossbeam::mem::epoch::participant::Participant

pub struct Participant {
    epoch: AtomicUsize,
    in_critical: AtomicUsize,
    garbage: UnsafeCell<Local>,
    pub active: AtomicBool,
    pub next: Atomic<ParticipantNode>,
}

Thread-local data for epoch participation.

Fields
epoch: AtomicUsize

The local epoch.

in_critical: AtomicUsize

Number of pending uses of epoch::pin(); keeping a count allows for reentrant use of epoch management.

garbage: UnsafeCell<Local>

Thread-local garbage tracking

active: AtomicBool

Is the thread still active? Becomes false when the thread exits. This is ultimately used to free Participant records.

next: Atomic<ParticipantNode>

The participant list is coded intrusively; here's the next pointer.

Methods

impl Participant

pub fn new() -> Participant

pub fn enter(&self) -> bool

Enter a critical section.

This method is reentrant, allowing for nested critical sections.

Returns true is this is the first entry on the stack (as opposed to a re-entrant call).

pub fn exit(&self)

Exit the current (nested) critical section.

pub unsafe fn reclaim<T>(&self, data: *mut T)

Begin the reclamation process for a piece of data.

pub fn try_collect(&self, guard: &Guard) -> bool

Attempt to collect garbage by moving the global epoch forward.

Returns true on success.

pub fn migrate_garbage(&self)

Move the current thread-local garbage into the global garbage bags.

pub fn garbage_size(&self) -> usize

How much garbage is this participant currently storing?

pub fn should_gc(&self) -> bool

Is this participant past its local GC threshhold?