Struct crossbeam::mem::epoch::participants::Participants

pub struct Participants {
    head: Atomic<ParticipantNode>,
}

Global, threadsafe list of threads participating in epoch management.

Fields

head

Methods

impl Participants

fn new() -> Participants

fn enroll(&self) -> *const Participant

Enroll a new thread in epoch management by adding a new Particpant record to the global list.

fn iter<'a>(&'a self, g: &'a Guard) -> Iter<'a>