Struct crossbeam::mem::epoch::Guard

#[must_use]
pub struct Guard {
    // some fields omitted
}

An RAII-style guard for pinning the current epoch.

A guard must be acquired before most operations on an Atomic pointer. On destruction, it unpins the epoch.

Methods

impl Guard

unsafe fn unlinked<T>(&self, val: Shared<T>)

Assert that the value is no longer reachable from a lock-free data structure and should be collected when sufficient epochs have passed.

fn migrate_garbage(&self)

Move the thread-local garbage into the global set of garbage.