Trait crossbeam::mem::ZerosValid

pub unsafe trait ZerosValid { }

Types for which mem::zeroed() is safe.

If a type T: ZerosValid, then a sequence of zeros the size of T must be a valid member of the type T.