Struct crossbeam::queue::SegQueue

pub struct SegQueue<T> { /* private fields */ }

An unbounded multi-producer multi-consumer queue.

This queue is implemented as a linked list of segments, where each segment is a small buffer that can hold a handful of elements. There is no limit to how many elements can be in the queue at a time. However, since segments need to be dynamically allocated as elements get pushed, this queue is somewhat slower than ArrayQueue.

Examples

use crossbeam_queue::SegQueue;

let q = SegQueue::new();

q.push('a');
q.push('b');

assert_eq!(q.pop(), Some('a'));
assert_eq!(q.pop(), Some('b'));
assert!(q.pop().is_none());

Implementations

impl<T> SegQueue<T>

pub const fn new() -> SegQueue<T>

Creates a new unbounded queue.

Examples

use crossbeam_queue::SegQueue;

let q = SegQueue::<i32>::new();

pub fn push(&self, value: T)

Pushes an element into the queue.

Examples

use crossbeam_queue::SegQueue;

let q = SegQueue::new();

q.push(10);
q.push(20);

pub fn pop(&self) -> Option<T>

Pops an element from the queue.

If the queue is empty, None is returned.

Examples

use crossbeam_queue::SegQueue;

let q = SegQueue::new();

q.push(10);
assert_eq!(q.pop(), Some(10));
assert!(q.pop().is_none());

pub fn is_empty(&self) -> bool

Returns true if the queue is empty.

Examples

use crossbeam_queue::SegQueue;

let q = SegQueue::new();

assert!(q.is_empty());
q.push(1);
assert!(!q.is_empty());

pub fn len(&self) -> usize

Returns the number of elements in the queue.

Examples

use crossbeam_queue::SegQueue;

let q = SegQueue::new();
assert_eq!(q.len(), 0);

q.push(10);
assert_eq!(q.len(), 1);

q.push(20);
assert_eq!(q.len(), 2);