Struct crossbeam::CachePadded

pub struct CachePadded<T> { /* fields omitted */ }

Pad T to the length of a cacheline.

Sometimes concurrent programming requires a piece of data to be padded out to the size of a cacheline to avoid "false sharing": cachelines being invalidated due to unrelated concurrent activity. Use the CachePadded type when you want to avoid cache locality.

At the moment, cache lines are assumed to be 32 * sizeof(usize) on all architectures.

Warning: the wrapped data is never dropped; move out using ptr::read if you need to run dtors.

Methods

impl<T: ZerosValid> CachePadded<T>

pub fn zeroed() -> CachePadded<T>

A const fn equivalent to mem::zeroed().

impl<T> CachePadded<T>

pub fn new(t: T) -> CachePadded<T>

Wrap t with cacheline padding.

Warning: the wrapped data is never dropped; move out using ptr:read if you need to run dtors.