Module crossbeam::sync

Thread synchronization primitives.

Parker, a thread parking primitive.

ShardedLock, a sharded reader-writer lock with fast concurrent reads.

WaitGroup, for synchronizing the beginning or end of some computation.

Structs

Parker

A thread parking primitive.

ShardedLock

A sharded reader-writer lock.

ShardedLockReadGuard

A guard used to release the shared read access of a ShardedLock when dropped.

ShardedLockWriteGuard

A guard used to release the exclusive write access of a ShardedLock when dropped.

Unparker

Unparks a thread parked by the associated Parker.

WaitGroup

Enables threads to synchronize the beginning or end of some computation.