Enum crossbeam::sync::chase_lev::Steal

pub enum Steal<T> {
    Empty,
    Abort,
    Data(T),
}

When stealing some data, this is an enumeration of the possible outcomes.

Variants

Empty

The deque was empty at the time of stealing

Abort

The stealer lost the race for stealing data, and a retry may return more data.

Data

The stealer has successfully stolen some data.