Struct crossbeam::sync::chase_lev::Stealer

pub struct Stealer<T> {
    // some fields omitted
}

The stealing half of the work-stealing deque. Stealers have access to the opposite end of the deque from the worker, and they only have access to the steal method.

Stealers can be cloned to have more than one handle active at a time.

Methods

impl<T> Stealer<T>

fn steal(&self) -> Steal<T>

Steals work off the end of the queue (opposite of the worker's end)