Struct crossbeam::sync::chase_lev::Worker

pub struct Worker<T> {
    // some fields omitted
}

Worker half of the work-stealing deque. This worker has exclusive access to one side of the deque, and uses push and try_pop method to manipulate it.

There may only be one worker per deque, and operations on the worker require mutable access to the worker itself.

Methods

impl<T> Worker<T>

fn push(&mut self, t: T)

Pushes data onto the front of this work queue.

fn try_pop(&mut self) -> Option<T>

Pops data off the front of the work queue, returning None on an empty queue.