Struct crossbeam::sync::ms_queue::MsQueue

pub struct MsQueue<T> {
    head: CachePadded<Atomic<Node<T>>>,
    tail: CachePadded<Atomic<Node<T>>>,
}

A Michael-Scott lock-free queue, with support for blocking pops.

Usable with any number of producers and consumers.

Fields

head

tail

Methods

impl<T> MsQueue<T>

fn new() -> MsQueue<T>

Create a new, empty queue.

fn push_internal(&self, guard: &Guard, onto: Shared<Node<T>>, n: Owned<Node<T>>) -> Result<(), Owned<Node<T>>>

Attempt to atomically place n into the next pointer of onto.

If unsuccessful, returns ownership of n, possibly updating the queue's tail pointer.

fn push(&self, t: T)

Add t to the back of the queue, possibly waking up threads blocked on pop.

fn pop_internal(&self, guard: &Guard) -> Result<Option<T>, ()>

fn is_empty(&self) -> bool

Check if this queue is empty.

fn try_pop(&self) -> Option<T>

Attempt to dequeue from the front.

Returns None if the queue is observed to be empty.

fn pop(&self) -> T

Dequeue an element from the front of the queue, blocking if the queue is empty.