Module crossbeam::epoch

Epoch-based memory reclamation.

See the crossbeam-epoch crate for more information.

Structs

Atomic
An atomic pointer that can be safely shared between threads.

Collector
An epoch-based garbage collector.

CompareAndSetError
The error returned on failed compare-and-set operation.

Guard
A guard that keeps the current thread pinned.

Handle
A handle to a garbage collector.

Owned
An owned heap-allocated object.

Shared
A pointer to an object protected by the epoch GC.

Traits

CompareAndSetOrdering
Memory orderings for compare-and-set operations.

Pointer
A trait for either Owned or Shared pointers.

Functions

default_collector
Returns the default handle associated with the current thread.

default_handle
Returns the default handle associated with the current thread.

is_pinned
Returns true if the current thread is pinned.

pin
Pins the current thread.

unprotected⚠
Returns a reference to a dummy guard that allows unprotected access to Atomics.