Struct crossbeam::mem::epoch::participant::Participant

pub struct Participant {
    epoch: AtomicUsize,
    in_critical: AtomicUsize,
    garbage: UnsafeCell<Local>,
    pub active: AtomicBool,
    pub next: Atomic<ParticipantNode>,
}

Thread-local data for epoch participation.

Fields

epoch

The local epoch.

in_critical

Number of pending uses of epoch::pin(); keeping a count allows for reentrant use of epoch management.

garbage

Thread-local garbage tracking

active

Is the thread still active? Becomes false when the thread exits. This is ultimately used to free Participant records.

next

The participant list is coded intrusively; here's the next pointer.

Methods

impl Participant

fn new() -> Participant

fn enter(&self) -> bool

Enter a critical section.

This method is reentrant, allowing for nested critical sections.

Returns true is this is the first entry on the stack (as opposed to a re-entrant call).

fn exit(&self)

Exit the current (nested) critical section.

unsafe fn reclaim<T>(&self, data: *mut T)

Begin the reclamation process for a piece of data.

fn try_collect(&self, guard: &Guard) -> bool

Attempt to collect garbage by moving the global epoch forward.

Returns true on success.

fn migrate_garbage(&self)

Move the current thread-local garbage into the global garbage bags.

fn garbage_size(&self) -> usize

How much garbage is this participant currently storing?

fn should_gc(&self) -> bool

Is this participant past its local GC threshhold?